feature: 입력,삭제,토글 기능 추가
