+++ src/components/Modal/AlarmModal.tsx
@@ -1,22 +1,80 @@
 import React from "react";
+import { useState } from "react";
 import styled from "styled-components";
 import { SubTitle2Typo } from "../../styles/Typography";
 import bellIcon from "../../../public/assets/Modal/notification-icon.svg";
 import NotificationRegister from "../Notification/NotificationRegister";
 import NotificationList from "../Notification/NotificationList";
 import NotificationHistory from "../Notification/NotificationHistory";
+import { notificationType } from "../Notification/NotificationType";
+
 
 interface ModalProps {
   closeModal: () => void;
 }
 
+const NotificationDummy : notificationType[]= [
+  {
+  id: 1,
+  name:"도지코인",
+  symbol:"DOGE",
+  volatility:30,
+  period:2,
+  direction:"RISE",
+  isAlertOn: "ON",
+  },
+  {
+  id: 2,
+  name:"봉크",
+  symbol:"BONK",
+  volatility:3,
+  period:60,
+  direction:"RISE",
+  isAlertOn: "ON",
+  },
+  {
+  id: 3,
+  name:"페페",
+  symbol:"PEPE",
+  volatility:3,
+  period:2,
+  direction:"FALL",
+  isAlertOn: "ON",
+  }
+];
+
+
 const AlarmModal: React.FC<ModalProps> = ({ closeModal }) => {
+  const [notifications, setNotifications] = useState(NotificationDummy);
+  const [nextId, setNextId] = useState(4);
+
   const handleOverlayClick = (e: React.MouseEvent) => {
     if (e.target === e.currentTarget) {
       closeModal();
     }
   };
 
+  const createNotifcation = (newNotification : Omit<notificationType, "id">) => {
+    const notificationWithId = {...newNotification, id : nextId};
+    setNextId((prevId) => prevId + 1); // 다음 id값 증가
+    setNotifications((prevNotifications)=>[...prevNotifications, notificationWithId]);
+  };
+
+  const toggleNotification = (id : number) => {
+    setNotifications((prevNotifications)=>
+        prevNotifications.map((notification)=>
+            notification.id === id
+                ? {...notification, isAlertOn: notification.isAlertOn === "ON" ? "OFF" : "ON" }
+                : notification
+        )
+    );
+};
+
+const deleteNotification = (id : number) => {
+    setNotifications((prevNotifications)=>
+            prevNotifications.filter((notification)=> notification.id !== id)
+    );
+};
 
   return (
     <ModalOverlay onClick={handleOverlayClick}>
@@ -27,10 +85,10 @@ const AlarmModal: React.FC<ModalProps> = ({ closeModal }) => {
         </FlexContainer>
         <NotificationContainer>
           <LeftSide>
-            <NotificationRegister></NotificationRegister>
-            <NotificationList></NotificationList>
+            <NotificationRegister createNotification={createNotifcation}/>
+            <NotificationList notifications={notifications} toggleNotification={toggleNotification} deleteNotification={deleteNotification}/>
           </LeftSide>
-          <DividerLine></DividerLine>
+          <DividerLine/>
           <RightSide>
             <NotificationHistory></NotificationHistory>
           </RightSide>
