+++ src/components/Notification/NotificationList.tsx
@@ -1,57 +1,15 @@
 import styled from "styled-components";
 import * as S from "../../styles/Typography.ts";
-import { useState } from "react";
 import NotificationCard from "./NotificationCard.tsx";
 import { notificationType } from "./NotificationType.ts";
 
-const NotificationDummy : notificationType[]= [
-    {
-    id: 1,
-    name:"도지코인",
-    symbol:"DOGE",
-    volatility:30,
-    period:2,
-    direction:"RISE",
-    isAlertOn: "ON",
-    },
-    {
-    id: 2,
-    name:"봉크",
-    symbol:"BONK",
-    volatility:3,
-    period:60,
-    direction:"RISE",
-    isAlertOn: "ON",
-    },
-    {
-    id: 3,
-    name:"페페",
-    symbol:"PEPE",
-    volatility:3,
-    period:2,
-    direction:"FALL",
-    isAlertOn: "ON",
-    }
-];
-
-const NotificationList = () => {
-    const [notifications, setNotifications] = useState(NotificationDummy);
-
-    const toggleNotification = (id : number) => {
-        setNotifications((prevNotifications)=>
-            prevNotifications.map((notification)=>
-                notification.id === id
-                    ? {...notification, isAlertOn: notification.isAlertOn === "ON" ? "OFF" : "ON" }
-                    : notification
-            )
-        );
-    };
-
-    const deleteNotification = (id : number) => {
-        setNotifications((prevNotifications)=>
-                prevNotifications.filter((notification)=> notification.id !== id)
-        );
-    };
+type NotificationListProps = {
+    notifications : notificationType[];
+    toggleNotification : (id : number)=>void;
+    deleteNotification : (id : number)=> void;
+}
+const NotificationList : React.FC<NotificationListProps> = ({notifications, toggleNotification, deleteNotification}) => {
+    
     return <Container>
         <S.SubTitle3Typo>등록된 알림 목록</S.SubTitle3Typo>
         <Content>
